✨ Add new CoreDomain and integrate into MainWindow
A new class, CoreDomain, has been added to encapsulate the functionality of combining and converting two strings. This class is injected into the MainWindow, where its method DoMagic is utilized in a button click event. Two textfield inputs have been added for users to input strings, and a label is used to display the output generated by the DoMagic method. These changes are done to make the code more modular and follow the SOLID principles.